01_chat/pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-chat"
version = "0.1.0"
description = "AI chat backend built with FastAPI."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.14,<2.0",
    "email-validator>=2.2,<3.0",
    "fastapi>=0.115,<1.0",
    "fastapi-users>=14.0,<15.0",
    "fastapi-users-db-sqlalchemy>=7.0,<8.0",
    "limits>=4.2,<5.0",
    "openai>=1.68,<2.0",
    "PyJWT>=2.10,<3.0",
    "pwdlib[argon2]>=0.2,<1.0",
    "pydantic-settings>=2.6,<3.0",
    "psycopg[binary]>=3.2,<4.0",
    "sse-starlette>=2.2,<3.0",
    "slowapi>=0.1.9,<1.0",
    "sqlalchemy>=2.0,<3.0",
    "uvicorn>=0.32,<1.0",
]

[project.optional-dependencies]
dev = [
    "httpx>=0.28,<1.0",
    "pytest>=8.3,<9.0",
    "ruff>=0.8",
    "mypy>=1.11",
]

[tool.pytest.ini_options]
pythonpath = ["src"]
testpaths = ["tests"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_unused_configs = true
mypy_path = "src"
packages = ["auditpy"]
exclude = ["build", "tests/golden"]
